django/main/apps/departments/serializers.py
from rest_framework import serializers

from main.apps.departments.models import Department, Company


class CompanySerializer(serializers.ModelSerializer):

    class Meta:
        model = Company
        fields = '__all__'


class DepartmentSerializer(serializers.ModelSerializer):

    class Meta:
        model = Department
        fields = '__all__'
