django/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
amqp==5.0.2               # via kombu
appnope==0.1.0            # via ipython
asgiref==3.2.10           # via django
autopep8==1.5.3           # via django-silk
backcall==0.2.0           # via ipython
billiard==3.6.3.0         # via celery
cairocffi==1.2.0          # via cairosvg, weasyprint
cairosvg==2.5.0           # via weasyprint
celery==5.0.2             # via -r requirements.in
certifi==2020.6.20        # via requests, sentry-sdk
cffi==1.14.3              # via cairocffi, weasyprint
chardet==3.0.4            # via requests
click-didyoumean==0.0.3   # via celery
click-repl==0.1.6         # via celery
click==7.1.2              # via celery, click-didyoumean, click-repl, uvicorn
cssselect2==0.4.1         # via cairosvg, weasyprint
decorator==4.4.2          # via ipython
defusedxml==0.6.0         # via cairosvg
django-braces==1.14.0     # via django-model-controller
django-cors-headers==3.5.0  # via -r requirements.in
django-crispy-forms==1.9.1  # via django-model-controller
django-environ==0.4.5     # via -r requirements.in
django-excel-tools==1.0.5  # via -r requirements.in
django-extensions==3.0.9  # via -r requirements.in
django-filter==2.4.0      # via -r requirements.in
django-model-controller==0.4.0  # via -r requirements.in
django-restql==0.11.0     # via -r requirements.in
django-sequences==2.5     # via -r requirements.in
django-silk==4.1.0        # via -r requirements.in
django==3.1.3             # via -r requirements.in, django-braces, django-cors-headers, django-filter, django-model-controller, django-restql, django-sequences, django-silk, djangorestframework, djangorestframework-simplejwt, model-bakery
djangorestframework-simplejwt==4.5.0  # via -r requirements.in
djangorestframework==3.12.2  # via -r requirements.in, django-restql, djangorestframework-simplejwt
et-xmlfile==1.0.1         # via openpyxl
factory-boy==3.1.0        # via -r requirements.in
faker==4.14.2             # via factory-boy
flake8==3.8.4             # via -r requirements.in
gprof2dot==2019.11.30     # via django-silk
gunicorn==20.0.4          # via -r requirements.in
h11==0.11.0               # via uvicorn
html5lib==1.1             # via weasyprint
httptools==0.1.1          # via -r requirements.in
idna==2.9                 # via requests
importlib-metadata==2.0.0  # via flake8, kombu, markdown
ipdb==0.13.4              # via -r requirements.in
ipython-genutils==0.2.0   # via traitlets
ipython==7.19.0           # via ipdb
jdcal==1.4.1              # via openpyxl
jedi==0.17.2              # via ipython
jinja2==2.11.2            # via django-silk
kombu==5.0.2              # via celery
markdown==3.3.3           # via -r requirements.in
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8
model-bakery==1.2.0       # via -r requirements.in
openpyxl==3.0.5           # via -r requirements.in
parso==0.7.1              # via jedi
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==7.2.0             # via -r requirements.in, cairosvg
prompt-toolkit==3.0.8     # via click-repl, ipython
psycopg2-binary==2.8.6    # via -r requirements.in
ptyprocess==0.6.0         # via pexpect
pycodestyle==2.6.0        # via autopep8, flake8
pycparser==2.20           # via cffi
pyflakes==2.2.0           # via flake8
pygments==2.6.1           # via django-silk, ipython
pyjwt==1.7.1              # via djangorestframework-simplejwt
pymupdf==1.17.7           # via -r requirements.in
pypeg2==2.15.2            # via django-restql
pyphen==0.10.0            # via weasyprint
python-dateutil==2.8.1    # via django-silk, faker
pytz==2020.1              # via -r requirements.in, celery, django, django-silk
redis==3.5.3              # via -r requirements.in
requests==2.24.0          # via django-silk
sentry-sdk==0.19.2        # via -r requirements.in
six==1.15.0               # via click-repl, django-braces, html5lib, python-dateutil
sqlparse==0.3.1           # via django, django-silk
tblib==1.7.0              # via -r requirements.in
text-unidecode==1.3       # via faker
tinycss2==1.1.0           # via cairosvg, cssselect2, weasyprint
toml==0.10.1              # via autopep8
traitlets==5.0.5          # via ipython
typing-extensions==3.7.4.3  # via uvicorn
unittest-xml-reporting==3.0.4  # via -r requirements.in
urllib3==1.25.9           # via requests, sentry-sdk
uvicorn==0.12.2           # via -r requirements.in
uvloop==0.14.0            # via -r requirements.in
vine==5.0.0               # via amqp, celery
wcwidth==0.2.5            # via prompt-toolkit
weasyprint==51            # via -r requirements.in
webencodings==0.5.1       # via cssselect2, html5lib, tinycss2
zipp==3.4.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
